Delete sneaky .DS_Store and rename everything
When I originally ran a Python script over the files it accidentally
renamed a .DS_Store file which is used for internal macOS stuff. The
gitignore missed it of course because it had the wrong name. Now it has
been deleted and all of the file names have been updated to stay
correct.